Add pre-commit framework to repo
This will eliminate the need for the CI/CD service to have to worry
about running tasks that would be better suited to run locally before
being pushed as part of a commit.

The GitHub actions 'Deploy' workflow has now been changed to just
'Test'. As of this writing, I do not intend on having GitHub actions
update the lambda if any related changes were made. Instead Terraform
will perform this update.