Integrate custom domain into the lambda
aws_apigatewayv2_domain_name uses a ACM cert resource but it has to
indirectly refer to the validation resource so that way Terraform does
not try and create a custom domain name with a cert not validated yet.

According to AWS's API Gateway documentation, API gateway manages the
creation of log groups of execution logging. While I can see the
convention used, and can probably Terraform the log group for now I am
ok skipping on its creation (the stage name is used as part of naming
the log group, so $default might be an issue). It seems I had an issue
with getting the custom domain working and giving the Api Gateway
cloudwatch permissions seemed to fix the issue.

No api mapping path is needed under the custom domain name because the
custom domain will always refer to the same api.