Remove auto merging deps from renovate
Considering this project is not a large system that can be exploited,
there is not much to gain from just updating dependencies over time.
This stemmed from finding out this morning that the Terraform lockfile
would be automatically updated and merged with updates. Not great
considering there have not been any Terraform tests written yet.

While this may change, currently I want to make sure the project will
remain in a reusable form even if I don't touch it for a considerable
amount of time. At least at this point in time, all parts do work with
said dependencies versions.