Correct examples, tests, and test descriptions
Majority if not all examples made use of double quotes when javascript
uses single quotes for strings (double quotes are used in the case of JSON
strings).

Both isValid and validate still had older examples where instead of JSON
as input, they had strings instead. As of current, it is expected that
parsed JSON will be the input to these functions. This extends to the
tests themselves because the descriptions were misleading in what the
actual input was to many of the calls for both functions.

Where possible, I tried to make sure the examples were a part of the
tests and to get the inputs as close as possible to what the examples
describe.