Add usage examples in README
Instead of hardcoding the lambda API URL into each example, I decided on
using a Terraform output variable to determine the API URL's value. This
was done to provide a small amount of mitigation from having any bots
scrape the curl commands and potentially hammer the API. Invoking the
API would require additionally to clone the repository, which isn't
hard per se but adds another step for a bot to perform.